Fence snaps endowments and permissions for stable/flask
